requirements.txt
appdirs==1.4.4
click==8.1.7
depthcharge-ms==0.1.0
natsort==8.4.0
numpy==1.26.3
pandas==2.1.4
psutil==5.9.7
PyGithub==2.1.1
pytorch-lightning==1.9.5
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.3.2
spectrum-utils==0.4.2
tensorboard==2.15.1
torch==2.1.2
tqdm==4.66.1